avito-xapi/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.7.0
pydantic>=2.10.0
pyjwt>=2.10.0
httpx>=0.28.0
python-multipart>=0.0.20

# Avito HTTP client (TLS impersonation)
curl-cffi>=0.7.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
